fix UI

+++ Cosmetic Store Management System/Views/CustomerPage.xaml.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -14,6 +14,7 @@ using Microsoft.UI.Xaml.Media;
 using Microsoft.UI.Xaml.Navigation;
 using Windows.Foundation;
 using Windows.Foundation.Collections;
+using Windows.Storage;
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
@@ -59,11 +60,36 @@ public sealed partial class CustomerPage : Page
             ViewModel.GoToPage(item.Page);
         }
     }
-    private void deleteButton_Click(object sender, RoutedEventArgs e)
+    private async void deleteButton_Click(object sender, RoutedEventArgs e)
     {
-        var button = sender as Button;
-        var customer = button?.DataContext as Customer;
-        ViewModel.dao.DeleteCustomer(customer.Phone);
-        ViewModel.LoadData();
+        var language = ApplicationData.Current.LocalSettings.Values["appLanguage"];
+        var dialog = new ContentDialog();
+
+        if (language.Equals("en-US"))
+        {
+            dialog.Title = "Delete Customer";
+            dialog.Content = "Are you sure you want to delete this customer?";
+            dialog.PrimaryButtonText = "No";
+            dialog.SecondaryButtonText = "Yes";
+        }
+        else
+        {
+            dialog.Title = "Xóa khách hàng";
+            dialog.Content = "Bạn có chắc chắn muốn xóa khách hàng này không?";
+            dialog.PrimaryButtonText = "Không";
+            dialog.SecondaryButtonText = "Có";
+        }
+
+        dialog.SecondaryButtonStyle = (Style)Application.Current.Resources["AccentButtonStyle"];
+        dialog.XamlRoot = this.Content.XamlRoot;
+        var result = await dialog.ShowAsync();
+
+        if (result == ContentDialogResult.Secondary)
+        {
+            var button = sender as Button;
+            var customer = button?.DataContext as Customer;
+            ViewModel.dao.DeleteCustomer(customer.Phone);
+            ViewModel.LoadData();
+        }
     }
 }
